[EAD-15] Refactored PathUtil to extend support for different file formats
